+++ stocal/algorithms.py
@@ -21,8 +21,9 @@
 """Stochastic simulation algorithms
 
 This module provides implementations of different stochastic simulation
-algorithms. All implementations have to implement the TrajectorySampler
-interface by subclassing this abstract base class.
+algorithms. All implementations have to implement the
+StochasticSimulationAlgorithm interface by subclassing this abstract
+base class.
 
 The module also provides several helper classes for general stochastic
 simulation algorithms. DependencyGraph is a generic species-dependency
@@ -245,7 +246,7 @@ class PriorityQueue(object):
             self._queue[key] = instances
 
 
-class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
+class StochasticSimulationAlgorithm(with_metaclass(abc.ABCMeta, object)):
     """Abstract base class for stochastic trajectory samplers.
 
     This is the interface for stochastic trajectory samplers, i.e.
@@ -259,9 +260,9 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
     >>> for transition in trajectory:
     ...     print trajectory.time, trajectory.state, transition
 
-    When implementing a novel TrajectorySampler, make sure to only
-    generate random numbers using the TrajectorySampler.rng random
-    number generator.
+    When implementing a novel StochasticSimulationAlgorithm, make sure
+    to only generate random numbers using the
+    StochasticSimulationAlgorithm.rng random number generator.
     """
     def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
         """Initialize the sampler for the given Process and state.
@@ -340,7 +341,8 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         Must return a triple where the first item is the time of the
         proposed transition, the second item the transition itself,
         and the last item a tuple of additional arguments that are
-        passed through to calls to TrajectorySampler.perform_transition.
+        passed through to calls to
+        StochasticSimulationAlgorithm.perform_transition.
         Time and transition have to be picked from the correct
         probability distributions.
         """
@@ -365,7 +367,7 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         for every rule of the process.
         If overwritten by a subclass, the signature can have additional
         arguments which are populated with the argument tuple returned
-        by TrajectorySampler.propose_potential_transition.
+        by StochasticSimulationAlgorithm.propose_potential_transition.
         """
         self.step += 1
         self.time = time
@@ -402,7 +404,8 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
         Consider to overwrite self.propose_potential_transition,
         self.is_applicable, self.perform_transition,
         self.reject_transition or self.has_reached_end in favour of
-        overloading __iter__ when implementing a TrajectorySampler.
+        overloading __iter__ when implementing a
+        StochasticSimulationAlgorithm.
         """
         while not self.has_reached_end():
             time, transition, args = self.propose_potential_transition()
@@ -420,7 +423,15 @@ class TrajectorySampler(with_metaclass(abc.ABCMeta, object)):
             self.time = self.tmax
 
 
-class DirectMethod(TrajectorySampler):
+class TrajectorySampler(StochasticSimulationAlgorithm):
+    """Deprecated. Identical to StochasticSimulationAlgorithm"""
+    def __init__(self, *args, **opts):
+        warnings.warn("Use StochasticSimulationAlgorithm instead",
+                      DeprecationWarning)
+        super(TrajectorySampler, self).__init__(*args, **opts)
+
+
+class DirectMethod(StochasticSimulationAlgorithm):
     """Implementation of Gillespie's direct method.
 
     The standard stochastic simulation algorithm, published in
@@ -435,7 +446,7 @@ class DirectMethod(TrajectorySampler):
     calculated for all transitions and time and occuring transition
     are drawn from the appropriate probability distributions.
 
-    See help(TrajectorySampler) for usage information.
+    See help(StochasticSimulationAlgorithm) for usage information.
     """
     def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
         if any(not isinstance(r, Reaction) for r in process.transitions):
@@ -505,7 +516,7 @@ class DirectMethod(TrajectorySampler):
             self.propensities.update_item(trans, propensity)
 
 
-class FirstReactionMethod(TrajectorySampler):
+class FirstReactionMethod(StochasticSimulationAlgorithm):
     """Implementation of Gillespie's first reaction method.
 
     A stochastic simulation algorithm, published in
@@ -514,7 +525,7 @@ class FirstReactionMethod(TrajectorySampler):
     FirstReactionMethod works with processes that feature deterministic
     transitions, i.e. Event's.
 
-    See help(TrajectorySampler) for usage information.
+    See help(StochasticSimulationAlgorithm) for usage information.
     """
     def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
         self._transitions = []
@@ -576,7 +587,7 @@ class NextReactionMethod(FirstReactionMethod):
     FirstReactionMethod whose runtime scales logarithmically with the
     number of reactions.
 
-    See help(TrajectorySampler) for usage information.
+    See help(StochasticSimulationAlgorithm) for usage information.
     """
     def __init__(self, process, state, t=0., tmax=float('inf'), steps=None, seed=None):
         self.dependency_graph = DependencyGraph()
@@ -646,7 +657,7 @@ class AndersonMethod(NextReactionMethod):
     This sampler correctly treats non-autonomous transitions, i.e.
     transitions with time dependent stochastic rates.
 
-    See help(TrajectorySampler) for usage information.
+    See help(StochasticSimulationAlgorithm) for usage information.
     """
     def add_transition(self, transition):
         """Add transition with own internal clock (T, P)"""