+++ stocal/examples/events.py
@@ -1,8 +1,8 @@
 """Event example
 
 stocal.Event's can be added to a processes definition just like
-Reactions. Process.trajectory returns an TrajectorySampler that
-can cope with deterministic transitions (e.g. FirstReactionMethod).
+Reactions. Process.trajectory returns an StochasticSimulationAlgorithm
+that can cope with deterministic transitions (e.g. FirstReactionMethod).
 Sampler selection and usage is entirely transparent to the user.
 """
 import stocal